requirements.txt
Django==5.1.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.7
Pillow==10.0.1
openpyxl==3.1.2
pandas==2.1.1
boto3==1.29.7
django-storages==1.14.2
whitenoise==6.5.0
